Moves serve action to terminal command executer
